+++ examples/template-hydrogen-default/src/components/DefaultSeo.server.jsx
@@ -1,4 +1,4 @@
-import {useShopQuery, Seo} from '@shopify/hydrogen';
+import {useShopQuery, Seo, useServerRequest} from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
 /**
@@ -14,9 +14,13 @@ export default function DefaultSeo() {
     cache: {maxAge: 60 * 60 * 12, staleWhileRevalidate: 60 * 60 * 12},
   });
 
+  const {url} = useServerRequest();
+
   return (
     <Seo
-      defaultPage={{
+      type="defaultSeo"
+      data={{
+        url,
         title: shopName,
         description: shopDescription,
       }}