+++ packages/hydrogen/src/components/Seo/Seo.client.tsx
@@ -5,102 +5,54 @@ import {HomePageSeo} from './HomePageSeo.client';
 import {ProductSeo} from './ProductSeo.client';
 import {CollectionSeo} from './CollectionSeo.client';
 import {PageSeo} from './PageSeo.client';
-import {TitleSeo} from './TitleSeo.client';
-import {DescriptionSeo} from './DescriptionSeo.client';
-import {TwitterSeo} from './TwitterSeo.client';
-import {ImageSeo} from './ImageSeo.client';
 
-import {checkInvalidOverride} from './utilities';
+import type {DefaultPage, HomePage, Product, Collection, Page} from './types';
 
-import type {
-  DefaultPage,
-  HomePage,
-  Product,
-  Collection,
-  Page,
-  Title,
-  Description,
-  Twitter,
-  Image,
-} from './types';
+type Props =
+  | {
+      type: 'defaultSeo';
+      data: DefaultPage;
+    }
+  | {
+      type: 'homepage';
+      data: HomePage;
+    }
+  | {
+      type: 'product';
+      data: Product;
+    }
+  | {
+      type: 'collection';
+      data: Collection;
+    }
+  | {
+      type: 'page';
+      data: Page;
+    };
 
-interface Props {
-  defaultPage?: DefaultPage;
-  homePage?: HomePage;
-  product?: Product;
-  collection?: Collection;
-  page?: Page;
-  title?: Title;
-  description?: Description;
-  twitter?: Partial<Twitter>;
-  image?: Partial<Image>;
-}
+/**
+ * The `Seo` component renders SEO information on a webpage.
+ */
+export function Seo({type, data}: Props) {
+  let SeoMarkup = null;
 
-const validOverrides = [
-  {
-    baseProp: 'defaultPage',
-    validProps: ['image'],
-  },
-  {
-    baseProp: 'homePage',
-    validProps: ['description', 'twitter', 'image'],
-  },
-  {
-    baseProp: 'product',
-    validProps: [],
-  },
-  {
-    baseProp: 'collection',
-    validProps: [],
-  },
-  {
-    baseProp: 'page',
-    validProps: ['image'],
-  },
-  {
-    baseProp: 'title',
-    validProps: [],
-  },
-  {
-    baseProp: 'description',
-    validProps: ['homePage'],
-  },
-  {
-    baseProp: 'twitter',
-    validProps: ['homePage'],
-  },
-  {
-    baseProp: 'image',
-    validProps: ['defaultPage', 'homePage', 'page'],
-  },
-];
+  switch (type) {
+    case 'defaultSeo':
+      SeoMarkup = <DefaultPageSeo {...(data as DefaultPage)} />;
+      break;
+    case 'homepage':
+      SeoMarkup = <HomePageSeo {...(data as HomePage)} />;
+      break;
+    case 'product':
+      SeoMarkup = <ProductSeo {...(data as Product)} />;
+      break;
+    case 'collection':
+      SeoMarkup = <CollectionSeo {...(data as Collection)} />;
+      break;
+    case 'page':
+      SeoMarkup = <PageSeo {...(data as Page)} />;
+      break;
+  }
 
-export function Seo(props: Props) {
-  const {
-    defaultPage,
-    homePage,
-    product,
-    collection,
-    page,
-    title,
-    description,
-    twitter,
-    image,
-  } = props;
-
-  checkInvalidOverride(props, validOverrides);
-
-  return (
-    <>
-      {defaultPage && <DefaultPageSeo {...defaultPage} />}
-      {homePage && <HomePageSeo {...homePage} />}
-      {product && <ProductSeo {...product} />}
-      {collection && <CollectionSeo {...collection} />}
-      {page && <PageSeo {...page} />}
-      {title && <TitleSeo title={title} />}
-      {description && <DescriptionSeo description={description} />}
-      {twitter && <TwitterSeo {...twitter} />}
-      {image && <ImageSeo {...image} />}
-    </>
-  );
+  return SeoMarkup;
 }